Cria serializer de deputados com campo para gastos
