+++ politicos/views.py
@@ -1,25 +1,9 @@
 from datetime import date
-from rest_framework import generics, serializers
+from rest_framework import generics
 
 from politicos.filters import DeputadoFilter
-from politicos.models import Deputado, GastoCotaParlamentar
-
-
-class GastoCotaParlamentarSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = GastoCotaParlamentar
-        fields = '__all__'
-
-
-class DeputadoSerializer(serializers.ModelSerializer):
-
-    class Meta:
-        model = Deputado
-        fields = [
-            'id', 'nome', 'partido', 'uf', 'id_legislatura',
-            'gastos'
-        ]
-        depth = 2
+from politicos.models import Deputado
+from politicos.serializers import DeputadoSerializer
 
 
 class DeputadoListView(generics.ListAPIView):
